+++ smart-contracts/contracts-common/concordium-contracts-common/src/types.rs
@@ -492,7 +492,6 @@ impl str::FromStr for Timestamp {
     type Err = ParseTimestampError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        use convert::TryInto;
         let datetime =
             chrono::DateTime::parse_from_rfc3339(s).map_err(ParseTimestampError::ParseError)?;
         let millis = datetime
@@ -2137,8 +2136,6 @@ mod serde_impl {
         type Err = ExchangeRateConversionError;
 
         fn from_str(s: &str) -> Result<Self, Self::Err> {
-            use convert::TryInto;
-
             let mut decimal = rust_decimal::Decimal::from_str_exact(s)?;
             decimal.normalize_assign();
             if decimal.is_zero() || decimal.is_sign_negative() {