+++ smart-contracts/contracts-common/concordium-contracts-common/CHANGELOG.md
@@ -7,14 +7,19 @@
   which ensures a valid length, and the unchecked method `new_unchecked`.
   - Migrate from `From`/`Into`: Use `new_unchecked` instead (if known to be
     valid length).
-- Make inner field in `OwnedParameter`/`Parameter` private, but add an `From`
+- Make inner field in `OwnedParameter`/`Parameter` private, but add a `From`
   implementation for getting the raw bytes.
-  - Migrate from `parameter.0`: use `Parameter.into()` instead (for both of the affected
+  - Migrate from `parameter.0`: use `parameter.into()` instead (for both of the affected
     types).
 - For `ModuleReference`, replace `AsRef<[u8;32]>` with `AsRef<[u8]>` and make
-  inner bytes public.
+  inner `bytes` public.
   - The change was necessary for internal reasons.
   - Migrate from `module_reference.as_ref()`: use `&module_reference.bytes` instead.
+- Replace `OwnedParameter::new` with `OwnedParameter::from_serial`, which also
+  ensures a valid length.
+  - Migrate from `new(x)`: Use `from_serial(x).unwrap()` (if known to be valid length).
+- Add an `empty` method for both `OwnedParameter` and `Parameter`.
+- Implement `Default` for `Parameter`.
 
 ## concordium-contracts-common 5.2.0 (2023-02-08)