+++ smart-contracts/contracts-common/concordium-contracts-common/CHANGELOG.md
@@ -6,6 +6,8 @@
   This allows using non-copyable (and non-clonable) types as map keys or set
   values.
 - Add the method `serial_for_smart_contracts` to `OwnedPolicy`, which serializes the policy for easy consumption by smart contracts.
+- Set minimum Rust version to 1.60.
+- Set Rust edition to 2021.
 
 ## concordium-contracts-common 5.3.1 (2023-04-12)